[support_report] rename file support.rb -> support_report.rb
[support_report] rename support.rb to support_report.rb
