+++ lib/appoptics_apm/support_report.rb
